requirements.txt
# AI Automation Suite - Requirements
# Python 3.11+ required

# Core Google API dependencies
google-api-python-client==2.100.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.1
gspread==5.10.0

# API and networking
requests==2.31.0
urllib3==2.0.7

# Environment and configuration
python-dotenv==1.0.0
PyYAML==6.0.1

# Image and video processing
Pillow==10.0.1
numpy==1.24.3

# Text processing
Unidecode==1.3.7

# Build tools (optional, for creating executables)
pyinstaller==5.13.2
pyperclip==1.8.2


# Testing (optional, for development)
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Linting and formatting (optional, for development)
ruff==0.1.6
black==23.11.0

# Type checking (optional, for development)
mypy==1.7.0
types-requests==2.31.0.10
types-PyYAML==6.0.12.12